feat(setup): bind the first-run routes and mint the token they need
SetupGate was written and tested and constructed nowhere, so the engine
could not set up a deployment at all: no token was ever minted, and both
routes answered the not-implemented fallback.

Open builds the gate and mints a token when the account count is zero,
writing it under the data directory at 0600. Without that the form is
unreachable, because the token is what opens the form that would ask for
it. A deployment that already has an account mints none.

The account and the deployment's first network settings travel in one
request. The settings are probed before anything is created, since
creating the account is what shuts the gate and there is no second pass
at the form. A host list that excludes the caller warns rather than
refuses: naming the DNS name a server will be reached under while
connected by address is the ordinary way to set one up, and no rule
separates it from the typo it resembles.

The first administrator's grant is core.GrantEveryShare rather than
grant rows built at the call site, and it runs inside the gate's lock:
an account that exists is one that can already reach the shares.

AccountCounter takes the account service's own width. It was int against
an int64 source, which is a narrowing conversion on the one value that
decides whether the window is open.

Eight mutations, all caught. Two serve tests named the setup route as
their example of an unbound one and now name events.

+++ go/engine/http/server/setup.go
@@ -45,8 +45,13 @@ var (
 )
 
 // AccountCounter is how the gate asks whether setup is finished.
+//
+// The width is the account service's own rather than a narrower one. This is
+// the value that decides whether the window is open, and a narrowing
+// conversion on it could wrap a populated deployment's count to zero and
+// reopen setup.
 type AccountCounter interface {
-	CountUsers(ctx context.Context) (int, error)
+	CountUsers(ctx context.Context) (int64, error)
 }
 
 // SetupGate guards first-administrator creation.
